refactor: remove ShareLinkActivity and integrate shared URL handling into AccountSelectActivity
